Fix: fix typeshare error

+++ backend/src/routes/message/paginate.rs
@@ -91,7 +91,7 @@ pub async fn route(
 
             let q = Message::find()
                 .filter(message::Column::ChatId.eq(limit.chat_id))
-                .limit(limit.limit.unwrap_or(MAX_PAGINATE_LIMIT));
+                .limit(limit.limit.unwrap_or(MAX_PAGINATE_LIMIT) as u64);
             let q = match (limit.order, limit.id) {
                 (MessagePaginateReqOrder::GT, None) => q.order_by_asc(message::Column::Id),
                 (MessagePaginateReqOrder::GT, Some(id)) => q
@@ -118,7 +118,7 @@ pub async fn route(
 
             let q = Message::find()
                 .filter(message::Column::ChatId.eq(range.chat_id))
-                .limit(MAX_PAGINATE_LIMIT)
+                .limit(MAX_PAGINATE_LIMIT as u64)
                 .filter(message::Column::Id.gt(range.lower).lt(range.upper));
             q
         }
